fix: change dx and dy types from uint32_t to int8_t for improved input handling

+++ shared/Network/Data/InputData.hpp
@@ -14,8 +14,8 @@
 struct InputData {
     HeaderData header;
     uint32_t entity;
-    uint32_t dx;
-    uint32_t dy;
+    int8_t dx;
+    int8_t dy;
     uint8_t shooting;
 };
